// packages/core/glue.d.ts
export namespace Glue42Core {

    export type LogLevel = "off" | "trace" | "debug" | "info" | "warn" | "error";

    /** Optional configuration object when initializing the Glue42 library. */
    export interface Config {
        /**
         * Application name. If not specified, the value depends on the hosting environment.
         * For the browser - `document.title + random number` (e.g., ClientList321333).
         * In Glue42 - `containerName + browserWindowName` (e.g., Internal.ClientList).
         */
        application?: string;

        /** Configurations for the Glue42 Gateway connection. */
        gateway?: GatewayConfig;

        /** Metrics configurations. */
        metrics?: boolean | MetricsConfig;

        /** Enable or disable the Contexts API. */
        contexts?: boolean;

        /** Enable or disable the Pub/Sub API. */
        bus?: boolean;

        /** Defines logging levels per output target. */
        logger?: LogLevel | LoggerConfig;

        /** Pass this to override the build-in logger and handle logging on your own */
        customLogger?: Glue42Core.CustomLogger;

        /**
         * Authentication can use one of the following flows:
         * - username/password;
         * - `token` - access tokens can be generated after successful login from the Auth Provider (e.g., Auth0);
         * - `gatewayToken` - Gateway tokens are time limited tokens generated by the Gateway after an explicit request. To generate one, use the `glue.connection.authToken()` method;
         * - `sspi` - using `sessionId` and authentication challenge callback;
         */
        auth?: Glue42Core.Auth | string;
    }
// ...
    export interface GlueCore {
        /** Connection library. */
        connection: Glue42Core.Connection.API;

        logger: Glue42Core.Logger.API;

        /** Interop library. */
        agm: Glue42Core.Interop.API;

        /** Interop library. */
        interop: Glue42Core.Interop.API;

        /** Pub/Sub library. */
        bus: Glue42Core.Bus.API;

        /** Metrics library. */
        metrics: Glue42Core.Metrics.API;

        /** Contexts library. */
        contexts: Glue42Core.Contexts.API;

        /** Brings up the Glue42 Desktop feedback dialog. */
        feedback: (info?: FeedbackInfo) => void;

        /** Info object containing versions of all included libraries and Glue42 itself. */
        info: object;

        /** Glue42 version. */
        version: string;

        /**
         * @ignore
         * Performance data
         */
        performance?: object;

        /**
         * @ignore
         * Config as passed from the user
         */
        userConfig?: Config;

        /**
         * @ignore
         * The config used by the library. This is a transformed and possibly extended version of the userConfig object
         */
        config?: any;

        /**
         * Disposes Glue42 API. This will remove all Interop methods and streams registered by the application.
         */
        done(): Promise<void>;
    }

    /**
     * Allows customizing the feedback form
     */
    export interface FeedbackInfo {
        /** Will be added to the description field in the feedback form */
        message?: string;
    }

    export namespace Logger {
        export interface API {

            /** Name of the logger. */
            name: string;

            /** Version of the Logging API. */
            version?: string;

            /**
             * Creates a new logger which is a sub-logger of the current one.
             * @param name Name for the sub-logger.
             */
            subLogger(name: string): Logger.API;

            /**
             * Sets or gets the current threshold level for publishing to a file.
             * @param level Logger level.
             */
            publishLevel(level?: LogLevel): LogLevel | undefined;

            /**
             * Sets or gets the current threshold level for writing to the console.
             * @param level Logger level.
             */
            consoleLevel(level?: LogLevel): LogLevel | undefined;

            /**
             * Logging method.
             * @param message Message to log.
             * @param level Logging level for the message.
             */
            log(message: string, level?: LogLevel): void;

            /**
             * Method for logging messages at "trace" level.
             * @param message Message to log.
             */
            trace(message: string): void;

            /**
             * Method for logging messages at "debug" level.
             * @param message Message to log.
             */
            debug(message: string): void;

            /**
             * Method for logging messages at "info" level.
             * @param message Message to log.
             */
            info(message: string): void;

            /**
             * Method for logging messages at "warn" level.
             * @param message Message to log.
             */
            warn(message: string): void;

            /**
             * Method for logging messages at "error" level.
             * @param message Message to log.
             */
            error(message: string): void;

            /**
             * Checks if the logger can publish a log level
             * @param level
             */
            canPublish(level: LogLevel): boolean;
        }

    }
// ...
    export namespace Bus {
        export interface API {
            /**
             * Publishes the provided data on a specific topic.
             * An optional object can be provided to publish a message to specific peers.
             *
             * @example
             * ```javascript
             * glue.bus.publish("prices", { RIC: "VOD.L", price: 21.2 })
             * ```
             * @param topic Topic on which to publish the message.
             * @param data Data to publish.
             * @param options An optional object with message options.
             */
            publish(
                topic: string,
                data: object,
                options?: MessageOptions
            ): void;

            /**
             * Subscribe for receiving data published on specific topic on the message bus. The provided callback will be invoked for each received message.
             * An optional object can be provided to receive messages published only by specific peers.
             *
             * Returns a `Promise` which resolves if the subscription is successful. The `Promise` resolves with a subscription object which can be used to unsubscribe and stop receiving messages.
             *
             * @example
             * ```javascript
             * glue.bus.subscribe(
             *     "prices",
             *     function (data, topic, source) {
             *         console.log(data, topic, source);
             *   })
             * ```
             * @param topic Topic to which to subscribe.
             * @param callback Function that will handle the received data.
             * @param options An optional object with message options.
             */
            subscribe(
                topic: string,
                callback: (data: object, topic: string, source: Glue42Core.Interop.Instance) => void,
                options?: MessageOptions
            ): Promise<Subscription>;
        }

        /**
         * If `routingKey` is provided when publishing a message, the message will be delivered only to peers that have provided the same routing key
         * or no routing key during subscription.
         * If `routingKey` is provided when subscribing for a topic, only messages that have been published with a matching routing key or no routing key
         * at all will be received by the subscriber.
         * If `target` is provided when publishing a message, the message will be delivered only to subscribers for which all properties of the target object
         * match the respective fields on the subscribers identity.
         * If `target` is provided when subscribing for a topic, the provided object will be compared to the identity of the publishers of the messages
         * and only if they match will the messages be received by the subscriber.
         */
        export interface MessageOptions {

            /** Routing key specified by the publisher/subscriber. */
            routingKey?: string;

            /** Target specified by the publisher/subscriber. */
            target?: object;
        }

        /**
         * Subscription object that can be used to unsubscribe from a topic and stop receiving data.
         */
        export interface Subscription {
            unsubscribe: () => Promise<void>;
        }
    }
// ...
}